Backport "Reuse getMappingOptionByTarget in ParameterTarget component"
