+++ enterprise/frontend/src/metabase-enterprise/sandboxes/containers/QuestionParameterTargetWidget.jsx
@@ -2,8 +2,9 @@
 import { Component } from "react";
 
 import { QuestionLoaderHOC } from "metabase/containers/QuestionLoader";
-import ParameterTargetWidget from "metabase/parameters/components/ParameterTargetWidget";
 import { getParameterMappingOptions } from "metabase/parameters/utils/mapping-options";
+
+import ParameterTargetWidget from "../components/ParameterTargetWidget";
 
 class QuestionParameterTargetWidget extends Component {
   render() {
@@ -11,7 +12,13 @@ class QuestionParameterTargetWidget extends Component {
     const mappingOptions = question
       ? getParameterMappingOptions(question, null, question.card())
       : [];
-    return <ParameterTargetWidget {...props} mappingOptions={mappingOptions} />;
+    return (
+      <ParameterTargetWidget
+        {...props}
+        question={question}
+        mappingOptions={mappingOptions}
+      />
+    );
   }
 }
