+++ frontend/src/metabase/dashboard/actions/auto-wire-parameters/utils.ts
@@ -2,13 +2,13 @@ import _ from "underscore";
 
 import { isActionDashCard } from "metabase/actions/utils";
 import { getExistingDashCards } from "metabase/dashboard/actions/utils";
-import { getMappingOptionByTarget } from "metabase/dashboard/components/DashCard/utils";
 import {
   isQuestionDashCard,
   isVirtualDashCard,
 } from "metabase/dashboard/utils";
 import {
   type ParameterMappingOption,
+  getMappingOptionByTarget,
   getParameterMappingOptions,
 } from "metabase/parameters/utils/mapping-options";
 import type Question from "metabase-lib/v1/Question";
@@ -80,7 +80,6 @@ export function getMatchingParameterOption(
 
   const matchedOption = getMappingOptionByTarget(
     mappingOptions,
-    targetDashcard,
     targetDimension,
     targetQuestion,
     parameter,
