+++ frontend/src/metabase/dashboard/components/DashCard/DashCardParameterMapper/DashCardCardParameterMapperButton.tsx
@@ -143,6 +143,7 @@ export const DashCardCardParameterMapperButton = ({
             }}
             target={target}
             mappingOptions={mappingOptions}
+            selectedMappingOption={selectedMappingOption}
           />
         }
       >
